% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae.R
\docType{data}
\name{adae}
\alias{adae}
\title{adae}
\format{
A data frame with 3211 rows and 7 columns:
"SUBJID"  "AEPT"    "AESOC"   "AESTDYI" "AEENDYI" "AESEVCD" "AECONT"
}
\usage{
adae
}
\description{
A dataset containing information related to adverse event.
}
\keyword{datasets}
