% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{adlb}
\format{
\subsection{\code{adlb}}{

A data frame with 125123 rows and 7 columns:
"SUBJID"   "LBTEST"   "LBBASE"   "LBSTRESN" "LBSTUNIT" "VISITDY"  "VISIT"
\describe{
\item{SUBJID}{subject ID}
\item{VISITDY}{visit day}
}
}
}
\usage{
adlb
}
\description{
adlb
}
\keyword{datasets}
