% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adls}
\alias{adls}
\title{adls}
\format{
\subsection{\code{adls}}{

A data frame with 29118 rows and 11 columns:
"SUBJID"   "VISITDY"  "VISIT"    "LSCAT"    "LSNEW"    "LSSLD"    "LSLD"
"LSSITE"   "LSNEWANY" "LSREADER" "LSTYPE"
\describe{
\item{SUBJID}{subject ID}
}
}
}
\usage{
adls
}
\description{
A dataset containing information related to visit and lesion.
}
\keyword{datasets}
