% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpm.R
\docType{data}
\name{adpm}
\alias{adpm}
\title{adpm}
\format{
A data frame with 14924 rows and 4 columns:
"SUBJID"  "VISIT"   "VISITDY" "WEIGHT"
}
\usage{
adpm
}
\description{
A dataset containing information related to visit day and weight.
}
\keyword{datasets}
