% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adrsp}
\alias{adrsp}
\title{adrsp}
\format{
\subsection{\code{adrsp}}{

A data frame with 10856 rows and 9 columns:
"SUBJID"   "RSRESP"   "VISITDY"  "VISIT"    "RSCONFYN"
"RSTRESP"  "RSNTRESP" "RSNEW"    "RSREADER"
\describe{
\item{SUBJID}{subject ID}
}
}
}
\usage{
adrsp
}
\description{
A dataset containing information related to visit.
}
\keyword{datasets}
