% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{adsl}
\format{
\subsection{\code{adsl}}{

A data frame with 935 rows and 19 columns:
"SUBJID"   "TRT"      "ATRT"     "PRSURG"   "DTHDY"    "DTH"
"PFSDYCR"  "PFSCR"    "LIVERMET"  "DIAGMONS" "AGE"      "SEX"
"B_WEIGHT" "B_HEIGHT" "RACE"     "B_ECOG"   "HISSUBTY" "B_METANM" "DIAGTYPE"
\describe{
\item{SUBJID}{subject ID}
}
}
}
\usage{
adsl
}
\description{
A dataset containing information related to death and treatment/control.
}
\keyword{datasets}
