% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomark}
\alias{biomark}
\title{biomark}
\format{
\subsection{\code{biomark}}{

A data frame with 935 rows and 19 columns:
"SUBJID"   "BMMTNM1"  "BMMTR1"   "BMMTNM2"  "BMMTR2"   "BMMTNM3"  "BMMTR3"
"BMMTNM4"  "BMMTR4"  "BMMTNM5"  "BMMTR5"   "BMMTNM6"  "BMMTR6"   "BMMTNM7"
"BMMTR7"   "BMMTNM15" "BMMTR15"  "BMMTNM16"  "BMMTR16"
\describe{
\item{SUBJID}{subject ID}
}
}
}
\usage{
biomark
}
\description{
A dataset containing information related to biomarker.
}
\keyword{datasets}
