% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square_test.R
\name{chi_square_test}
\alias{chi_square_test}
\title{To test whether the marker is prognostic with respect to the death outcome.}
\usage{
chi_square_test(x)
}
\arguments{
\item{x}{a data frame.}
}
\value{
A contingency table of death vs. biomarker, a p-value indicating
whether the marker is prognostic with respect to the death outcome.
}
\description{
To test whether the marker is prognostic with respect to the death outcome.
}
