% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_KM_plot.R
\name{create_KM_plot}
\alias{create_KM_plot}
\title{To create a KM plot for survival analysis}
\usage{
create_KM_plot(data)
}
\arguments{
\item{data}{the database table with biomarker and treatment}
}
\value{
a survival plot
}
\description{
To create a KM plot for survival analysis
}
