% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survival_info.R
\name{create_survival_info}
\alias{create_survival_info}
\title{To create information from survival analysis}
\usage{
create_survival_info(data)
}
\arguments{
\item{data}{the database table with biomarker and treatment}
}
\value{
a survival plot
}
\description{
To create information from survival analysis
}
