% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_mutant.R
\name{determine_mutant}
\alias{determine_mutant}
\title{Determine whether a patient is mutant or wild-type.}
\usage{
determine_mutant(x, x_vars)
}
\arguments{
\item{x}{a data frame.}

\item{x_vars}{the variables in the dataframe referencing exon mutations,
can be taken in as a vector.}
}
\value{
The original data frame with a new variable "marker" indicating
whether the subject is mutant, wild-type, or other.
}
\description{
Determine whether a patient is mutant or wild-type.
}
