% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam.R
\name{fit_gam}
\alias{fit_gam}
\title{To model a generalized additive model}
\usage{
fit_gam(x)
}
\arguments{
\item{x}{the database table with biomarker and treatment}
}
\value{
a model summary
}
\description{
To model a generalized additive model
}
