% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gbm.R
\name{fit_gbm}
\alias{fit_gbm}
\title{To model a generalized boosted regression model}
\usage{
fit_gbm(x, x_vars)
}
\arguments{
\item{x}{the database table with biomarker and treatment}

\item{x_vars}{the predictors of interest, can be a vector of predictors}
}
\value{
a model summary
}
\description{
To model a generalized boosted regression model
}
