% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logistic.R
\name{fit_logistic}
\alias{fit_logistic}
\title{To model a logistic regression model on the joint effect
of biomarker and treatment}
\usage{
fit_logistic(x)
}
\arguments{
\item{x}{the database table with biomarker and treatment}
}
\value{
a model summary
}
\description{
To model a logistic regression model on the joint effect
of biomarker and treatment
}
