% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutant_cnt.R
\name{plot_mutant_cnt}
\alias{plot_mutant_cnt}
\title{Plot Barplot of Number of Mutants}
\usage{
plot_mutant_cnt(x, x_var = "marker")
}
\arguments{
\item{x}{a data frame.}

\item{x_var}{the variable in the dataframe indicating
whether a subject is mutant, wild-type, or unknown. Default is "marker".}
}
\value{
A ggplot object showing the number of total mutant, wild-type,
and unknown KRAS counts per patient.
}
\description{
Plot Barplot of Number of Mutants
}
