% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{adae}
\format{
\subsection{\code{adae}}{

A data frame with 3211 rows and 7 columns:
"SUBJID"  "AEPT"    "AESOC"   "AESTDYI" "AEENDYI" "AESEVCD" "AECONT"
\describe{
\item{SUBJID}{subject ID}
\item{AEPT}{AEPT}
\item{AESEVCD}{severity code}
}
}
}
\usage{
adae
}
\description{
adae
}
\keyword{datasets}
