% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlb.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{adlb}
\format{
A data frame with 125123 rows and 7 columns:
"SUBJID"   "LBTEST"   "LBBASE"   "LBSTRESN" "LBSTUNIT" "VISITDY"  "VISIT"
}
\usage{
adlb
}
\description{
A dataset containing information related to visit.
}
\keyword{datasets}
