% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\docType{data}
\name{adls}
\alias{adls}
\title{adls}
\format{
A data frame with 29118 rows and 11 columns:
"SUBJID"   "VISITDY"  "VISIT"    "LSCAT"    "LSNEW"    "LSSLD"    "LSLD"
"LSSITE"   "LSNEWANY" "LSREADER" "LSTYPE"
}
\usage{
adls
}
\description{
A dataset containing information related to visit and lesion.
}
\keyword{datasets}
