% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adpm}
\alias{adpm}
\title{adpm}
\format{
\subsection{\code{adpm}}{

A data frame with 14924 rows and 4 columns:
"SUBJID"  "VISIT"   "VISITDY" "WEIGHT"
\describe{
\item{SUBJID}{subject ID}
}
}
}
\usage{
adpm
}
\description{
A dataset containing information related to visit day and weight.
}
\keyword{datasets}
